Leave in _id and __v on lean queries
The code before was trying to strip out these unneeded fields from lean queries so they wouldn't be seen on the client. The problem was that the code was stripping _id while mongoose was still trying to use it to insert populated fields. So population broke. I considered a few ways of stripping the fields and then decided that it's not worth it at this point and it's fine to send them to the client :P