API v2! (#47)
Already? Resource.js looks like a better library in a few ways than `express-restify-mongoose`. And this will be an easier path to getting this layer to play nicely with some things needed for `mongoose-authz`. v1 and v2 are pretty similar. Probably some odds and ends to bring to parity, but wanted to get this out there sooner, rather than later.