User perms merge (#51)
* Have the base schema install the authz plugin if it detects a permissions object

This way the authz plugin gets installed in a single place. So if we want to set some options on the plugin, that can be done centrally.

* Prefetching the acting user's memberships

The org's the user is a part of are necessary for almost every authorization check. So get it once at the start of the request and pass it as part of the authPayload.

* Make sure permissions are serialized with mongoose documents

This makes sure that the permissions that authz adds to documents get serialized

* Upgrading mongoose-authz

Gets the latest race condition fix. Also specifies which un-protectable methods are allowed

* Permissions and Tests for the User Model

This is one of the gnarlier ones, and will probably evolve over time.